+++ crates/astria-cli/Cargo.toml
@@ -25,5 +25,4 @@ features = ["http"]
 
 [dev-dependencies]
 assert_cmd = "2.0.12"
-once_cell = { workspace = true }
-tempfile = { workspace = true }
+test-utils = { path = "./test-utils" }
