+++ crates/astria-cli/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod cli;
 pub mod commands;
-#[cfg(test)]
-pub(crate) mod test_utils;
 pub mod types;
